Moving group sampler to its own
