+++ include/epiworld/epiworld.hpp
@@ -69,6 +69,9 @@ namespace epiworld {
 
     #include "agentssample-bones.hpp"
 
+    #include "groupsampler-bones.hpp"
+    #include "groupsampler-meat.hpp"
+
     #include "models/models.hpp"
 
 }
